Closes #122
The default `robots.txt` now disallows all when the environment is NOT set to `'production'`.
This will not have an effect for existing installs (where the SEO settings have been saved).
To manually update your `robots.txt`, replace the line `{% if craft.app.config.general.devMode %}` with `{% if craft.app.config.env != 'production' %}`.